fix: remove old telemetry
