+++ .github/workflows/release-dev.yaml
@@ -52,6 +52,5 @@ jobs:
           args: release --rm-dist -f .goreleaser-dev.yml
         env:
           GITHUB_TOKEN: ${{ secrets.CI_BOT_TOKEN }}
-          TELEMETRY_TOKEN: ${{ secrets.TELEMETRY_TOKEN }}
           # Your GoReleaser Pro key, if you are using the 'goreleaser-pro' distribution
           # GORELEASER_KEY: ${{ secrets.GORELEASER_KEY }}
