+++ .goreleaser.yml
@@ -15,7 +15,6 @@ builds:
       - -X main.commit={{.Commit}}
       - -X main.date={{.Date}}
       - -X main.builtBy=goreleaser
-      - -X github.com/kubeshop/testkube/pkg/telemetry.telemetryToken={{.Env.TELEMETRY_TOKEN}}
 archives:
   - replacements:
       darwin: macOS
