+++ cmd/kubectl-testkube/commands/root.go
@@ -10,7 +10,6 @@ import (
 	"github.com/kubeshop/testkube/cmd/kubectl-testkube/config"
 	"github.com/kubeshop/testkube/pkg/analytics"
 	apiclient "github.com/kubeshop/testkube/pkg/api/v1/client"
-	"github.com/kubeshop/testkube/pkg/telemetry"
 	"github.com/kubeshop/testkube/pkg/ui"
 )
 
@@ -57,7 +56,6 @@ var RootCmd = &cobra.Command{
 
 		if analyticsEnabled {
 			ui.Debug("collecting anonymous analytics data, you can disable it by calling `kubectl testkube anlytics disable`")
-			telemetry.CollectAnonymousCmdInfo()
 			analytics.SendAnonymouscmdInfo()
 		}
 	},
